custom_nodes/MemedeckComfyNodes/requirements.txt
requests
numpy
torch
Pillow
opencv-python
torchvision